model_training/requirements.txt
# Core ML libraries
tensorflow>=2.13.0
tensorflow-datasets>=4.9.0
numpy>=1.21.0
pandas>=1.3.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Data science utilities
scikit-learn>=1.0.0
scipy>=1.7.0

# Development tools
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.6.0

# Configuration and utilities
PyYAML>=5.4.0
tqdm>=4.62.0
argparse

# Optional: Experiment tracking
# mlflow>=1.20.0
# wandb>=0.12.0

# Optional: Advanced visualization
# tensorboard>=2.7.0
# plotly>=5.3.0

# Optional: Data augmentation
# albumentations>=1.1.0
# imgaug>=0.4.0